refactor(components): split grid roles
