+++ ShipCoreFramework/src/Data/Scripts/ShipCoreFramework/Shared/Components/GridComponent.cs
@@ -16,9 +16,6 @@ namespace ShipCoreFramework
         private readonly object _shipControllersLock = new object();
         private readonly List<IMySlimBlock> _blocks = new List<IMySlimBlock>();
         private readonly List<IMyShipController> _shipControllers = new List<IMyShipController>();
-        private ConcurrentDictionary<BlockLimit, LimitBucket> _limits = new ConcurrentDictionary<BlockLimit, LimitBucket>();
-        internal ConcurrentDictionary<BlockLimit, LimitBucket> Limits => _limits;
-
         internal int BlockCount
         {
             get
@@ -30,14 +27,6 @@ namespace ShipCoreFramework
 
         internal readonly ConcurrentDictionary<IMyCubeBlock, CoreComponent> CoreDictionary =
             new ConcurrentDictionary<IMyCubeBlock, CoreComponent>();
-
-        internal readonly ConcurrentDictionary<IMyCubeBlock, BeaconComponent> BeaconDictionary =
-            new ConcurrentDictionary<IMyCubeBlock, BeaconComponent>();
-
-        private readonly ConcurrentDictionary<IMyCubeBlock, UpgradeModuleComponent> _upgradeModuleDictionary =
-            new ConcurrentDictionary<IMyCubeBlock, UpgradeModuleComponent>();
-
-        private readonly ConcurrentDictionary<long, byte> _trackedConnectorIds = new ConcurrentDictionary<long, byte>();
 
         private GroupComponent GroupComponent
         {
@@ -86,12 +75,6 @@ namespace ShipCoreFramework
             foreach (var otherBlock in otherBlocks) BlockAddedInternal(otherBlock);
         }
 
-        internal void PublishLimitsSnapshot(ConcurrentDictionary<BlockLimit, LimitBucket> limits)
-        {
-            System.Threading.Interlocked.Exchange(ref _limits,
-                limits ?? new ConcurrentDictionary<BlockLimit, LimitBucket>());
-        }
-
         internal List<IMyShipController> GetShipControllersCopy()
         {
             lock (_shipControllersLock)
@@ -122,7 +105,7 @@ namespace ShipCoreFramework
         private void GridMarkedForClose(IngameIMyEntity entity)
         {
             if (entity != Grid) return;
-            LimitsNexusSync.NotifyLocalGridClose();
+            if (Session.IsServer) NotifyLocalGridCloseAuthoritative();
         }
     }
 }
